+++ src/components/seekbar-playback-container/seekbar-playback-container.js
@@ -62,18 +62,21 @@ class SeekBarPlaybackContainer extends BaseComponent {
   render(): React$Element<any> {
     return (
       <SeekBarControl
-        playerElement={document.getElementById(this.config.targetId)}
+        player={this.props.player}
+        playerElement={this.props.playerContainer}
         showFramePreview={this.props.showFramePreview}
         showTimeBubble={this.props.showTimeBubble}
         changeCurrentTime={time => this.player.currentTime = time}
         playerPoster={this.props.poster}
         updateSeekbarDraggingStatus={data => this.props.updateSeekbarDraggingStatus(data)}
         updateCurrentTime={data => this.props.updateCurrentTime(data)}
-
         currentTime={this.props.currentTime}
         duration={this.props.duration}
         isDraggingActive={this.props.isDraggingActive}
         isMobile={this.props.isMobile}
+        thumbsSprite={this.props.config.thumbsSprite}
+        thumbsSlices={this.props.config.thumbsSlices}
+        thumbsWidth={this.props.config.thumbsWidth}
       />
     )
   }